refactor: Migrate all commands for their own files into commands folder.
