+++ src/libs/utils.ts
@@ -1,10 +1,10 @@
-import { existsSync, mkdirSync } from "fs";
-import { readFile, writeFile } from "fs/promises";
-import { dirname } from "path";
+import { existsSync, mkdirSync } from 'fs';
+import { readFile, writeFile } from 'fs/promises';
+import { dirname } from 'path';
 
-import yaml from "yaml";
-import chalk from "chalk";
-import log from "log-update";
+import yaml from 'yaml';
+import chalk from 'chalk';
+import log from 'log-update';
 
 interface IConfigFile {
     key: string;
@@ -14,26 +14,26 @@ interface IConfigFile {
 }
 
 interface IConfigFileLoaded {
-    ":api_key": string;
-    ":password": string;
-    ":theme_id": number;
-    ":preview_url": string;
+    ':api_key': string;
+    ':password': string;
+    ':theme_id': number;
+    ':preview_url': string;
 }
 
 export function saveConfigFile({ key, password, themeId, previewUrl }: IConfigFile) {
     const fileDataAsObject: IConfigFileLoaded = {
-        ":api_key": key,
-        ":password": password,
-        ":theme_id": themeId,
-        ":preview_url": previewUrl,
+        ':api_key': key,
+        ':password': password,
+        ':theme_id': themeId,
+        ':preview_url': previewUrl,
     };
 
     const configFileData = yaml.stringify(fileDataAsObject);
 
-    return writeFile("config.yml", configFileData)
+    return writeFile('config.yml', configFileData)
         .then(() => ({
             success: true,
-            message: "Configuration file created with success.",
+            message: 'Configuration file created with success.',
         }))
         .catch((error) => ({
             success: false,
@@ -42,11 +42,11 @@ export function saveConfigFile({ key, password, themeId, previewUrl }: IConfigFi
 }
 
 export async function loadConfigFile() {
-    return readFile("config.yml", { encoding: "utf8" })
+    return readFile('config.yml', { encoding: 'utf8' })
         .then((data) => {
             const config = yaml.parse(data) as IConfigFileLoaded;
 
-            const { ":api_key": key, ":password": password, ":theme_id": themeId, ":preview_url": previewUrl } = config;
+            const { ':api_key': key, ':password': password, ':theme_id': themeId, ':preview_url': previewUrl } = config;
 
             return {
                 success: true,
@@ -83,36 +83,36 @@ export async function saveAssetFile(path: string, data: Buffer) {
         }));
 }
 
-type LogMessageType = "info" | "pending" | "success" | "warning" | "error";
+type LogMessageType = 'info' | 'pending' | 'success' | 'warning' | 'error';
 
 export function logMessage(type: LogMessageType, message: string, done: boolean = false) {
     let color;
     let label;
 
     switch (type) {
-        case "pending":
-            color = "blue";
-            label = "Processing";
+        case 'pending':
+            color = 'blue';
+            label = 'Processing';
 
             break;
-        case "success":
-            color = "green";
-            label = "Complete";
+        case 'success':
+            color = 'green';
+            label = 'Complete';
 
             break;
-        case "warning":
-            color = "yellow";
-            label = "Warn";
+        case 'warning':
+            color = 'yellow';
+            label = 'Warn';
 
             break;
-        case "error":
-            color = "red";
-            label = "Fail";
+        case 'error':
+            color = 'red';
+            label = 'Fail';
 
             break;
         default:
-            color = "";
-            label = "";
+            color = '';
+            label = '';
 
             break;
     }
